requirements.txt
torch>=2.0.0
numpy>=1.24.0
mcschematic>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
tensorboard>=2.13.0
nbtlib>=2.0.0
